# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Castor Orçamento API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /produtos:
    get:
      operationId: listProdutos
      tags: [produtos]
      summary: Listar todos os produtos
      parameters:
        - name: categoria
          in: query
          required: false
          schema:
            type: string
        - name: limite
          in: query
          required: false
          schema:
            type: number
      responses:
        "200":
          description: Lista de produtos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Produto"
  /produtos/buscar:
    get:
      operationId: buscarProdutos
      tags: [produtos]
      summary: Buscar produtos por nome
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: categoria
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Produtos encontrados
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Produto"
...
  /produtos/{id}:
    get:
      operationId: getProduto
      tags: [produtos]
      summary: Buscar produto por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Produto encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Produto"
        "404":
          description: Produto não encontrado

  # ── Orçamento ─────────────────────────────────────────────────────────────
  /orcamento:
    post:
      operationId: gerarOrcamento
      tags: [orcamento]
      summary: Gerar orçamento formatado
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrcamentoInput"
      responses:
        "200":
          description: Orçamento gerado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrcamentoOutput"
  /orcamento/salvar:
    post:
      operationId: salvarOrcamento
      tags: [orcamento]
      summary: Salvar orçamento no banco de dados
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SalvarOrcamentoInput"
      responses:
        "200":
          description: Orçamento salvo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SalvarOrcamentoOutput"
  /orcamento/historico:
    get:
      operationId: historicoOrcamentos
      tags: [orcamento]
      summary: Listar histórico de orçamentos
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: number
      responses:
        "200":
          description: Lista de orçamentos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HistoricoItem"

  # ── Crawler ───────────────────────────────────────────────────────────────
...
  /entrada-estoque/produtos/buscar:
    get:
      operationId: buscarProdutosEstoque
      tags: [entrada-estoque]
      summary: Buscar produtos para correlação de estoque
      parameters:
        - $ref: "#/components/parameters/SessionToken"
        - name: q
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Produtos encontrados
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Produto"

  # ── Diagnóstico ───────────────────────────────────────────────────────────
...
components:
  parameters:
    SessionToken:
      name: x-session-token
      in: header
      required: false
      schema:
        type: string
  schemas:
...
    Produto:
      type: object
      properties:
        id:
          type: number
        nome:
          type: string
        sku:
          type: string
        preco:
          type: string
        precoPix:
          type: string
        parcelamento:
          type: string
        medidas:
          type: string
        altura:
          type: string
        categoria:
          type: string
        imagem:
          type: string
        link:
          type: string
        criadoEm:
          type: string
      required: [id, nome, categoria]
    OrcamentoInput:
      type: object
      properties:
        cliente:
          type: string
        whatsapp:
          type: string
        produtoIds:
          type: array
          items:
            type: number
        observacoes:
          type: string
        descontoPix:
          type: number
      required: [cliente, produtoIds]
    OrcamentoOutput:
      type: object
      properties:
        texto:
          type: string
        totalPrecoBase:
          type: string
        totalPix:
          type: string
        totalPrazo:
          type: string
        parcela12:
          type: string
        descontoAplicado:
          type: string
        descontoPercentual:
          type: number
        produtos:
          type: array
          items:
            $ref: "#/components/schemas/Produto"
      required: [texto, totalPix, totalPrazo, produtos]
    SalvarOrcamentoInput:
      type: object
      properties:
        cliente:
          type: string
        whatsapp:
          type: string
        produtosJson:
          type: array
          items:
            $ref: "#/components/schemas/Produto"
        observacoes:
          type: string
        descontoPix:
          type: number
        totalPix:
          type: string
        totalPrazo:
          type: string
        texto:
          type: string
        vendedor:
          type: string
        precoBaseTotal:
          type: string
        descontoAplicado:
          type: string
      required: [cliente, texto]
    SalvarOrcamentoOutput:
      type: object
      properties:
        id:
          type: number
        mensagem:
          type: string
      required: [id, mensagem]
    HistoricoItem:
      type: object
      properties:
        id:
          type: number
        cliente:
          type: string
        whatsapp:
          type: string
        status:
          type: string
        vendedor:
          type: string
        totalPix:
          type: string
        totalPrazo:
          type: string
        descontoPix:
          type: number
        observacoes:
          type: string
        texto:
          type: string
        produtosJson:
          type: array
          items:
            $ref: "#/components/schemas/Produto"
        criadoEm:
          type: string
      required: [id, cliente, status, texto]
...